Break methods into separate module
